MolTrack: Update batch-related endpoints to perform actions using corporate_batch_id

+++ app/crud/batches.py
@@ -3,14 +3,14 @@ from sqlalchemy import and_
 from sqlalchemy.orm import Session
 from sqlalchemy.orm import joinedload
 from app.crud.properties import enrich_properties
-from app import models
+from app import crud, models
 
 
 def enrich_batch(batch: models.Batch) -> models.BatchResponse:
     return models.BatchResponse(**batch.dict(), properties=enrich_properties(batch, "batch_details", "batch_id"))
 
 
-def get_batch(db: Session, corporate_batch_id: str):
+def get_batch(db: Session, corporate_batch_id: str, enrich_batch: bool = True):
     batch = (
         db.query(models.Batch)
         .join(models.Batch.batch_details)
@@ -25,7 +25,7 @@ def get_batch(db: Session, corporate_batch_id: str):
     if not batch:
         return None
 
-    return enrich_batch(batch)
+    return enrich_batch(batch) if enrich_batch else batch
 
 
 def get_batches(db: Session, skip: int = 0, limit: int = 100):
@@ -37,13 +37,15 @@ def get_batches_by_compound(db: Session, compound_id: int, skip: int = 0, limit:
     return db.query(models.Batch).filter(models.Batch.compound_id == compound_id).offset(skip).limit(limit).all()
 
 
-def delete_batch(db: Session, batch_id: int):
-    db_batch = db.get(models.Batch, batch_id)
+def delete_batch(db: Session, corporate_batch_id: str):
+    db_batch = get_batch(db, corporate_batch_id=corporate_batch_id, enrich_batch=False)
     if db_batch is None:
         raise HTTPException(status_code=404, detail="Batch not found")
 
-    db.query(models.BatchDetail).filter(models.BatchDetail.batch_id == batch_id).delete(synchronize_session=False)
-    db.query(models.BatchAddition).filter(models.BatchAddition.batch_id == batch_id).delete(synchronize_session=False)
+    assay_results = crud.get_all_assay_results_for_batch(db, db_batch.id)
+    if assay_results:
+        raise HTTPException(status_code=400, detail="Batch has dependent assay results")
+
     db.delete(db_batch)
     db.commit()
     return db_batch
